feat: adding deltalog stream reader and writer
