+++ internal/storage/serde.go
@@ -17,26 +17,18 @@
 package storage
 
 import (
-	"bytes"
-	"encoding/binary"
 	"fmt"
 	"io"
 	"math"
-	"sort"
-	"strconv"
 
 	"github.com/apache/arrow/go/v12/arrow"
 	"github.com/apache/arrow/go/v12/arrow/array"
-	"github.com/apache/arrow/go/v12/arrow/memory"
 	"github.com/apache/arrow/go/v12/parquet"
 	"github.com/apache/arrow/go/v12/parquet/pqarrow"
-	"github.com/cockroachdb/errors"
 	"github.com/golang/protobuf/proto"
 
 	"github.com/milvus-io/milvus-proto/go-api/v2/schemapb"
 	"github.com/milvus-io/milvus/pkg/common"
-	"github.com/milvus-io/milvus/pkg/util/metautil"
-	"github.com/milvus-io/milvus/pkg/util/typeutil"
 )
 
 type Record interface {
@@ -87,115 +79,6 @@ func (r *compositeRecord) Release() {
 
 func (r *compositeRecord) Schema() map[FieldID]schemapb.DataType {
 	return r.schema
-}
-
-var _ RecordReader = (*compositeRecordReader)(nil)
-
-type compositeRecordReader struct {
-	blobs [][]*Blob
-
-	blobPos int
-	rrs     []array.RecordReader
-	closers []func()
-	fields  []FieldID
-
-	r compositeRecord
-}
-
-func (crr *compositeRecordReader) iterateNextBatch() error {
-	if crr.closers != nil {
-		for _, close := range crr.closers {
-			if close != nil {
-				close()
-			}
-		}
-	}
-	crr.blobPos++
-	if crr.blobPos >= len(crr.blobs[0]) {
-		return io.EOF
-	}
-
-	for i, b := range crr.blobs {
-		reader, err := NewBinlogReader(b[crr.blobPos].Value)
-		if err != nil {
-			return err
-		}
-
-		crr.fields[i] = reader.FieldID
-		// TODO: assert schema being the same in every blobs
-		crr.r.schema[reader.FieldID] = reader.PayloadDataType
-		er, err := reader.NextEventReader()
-		if err != nil {
-			return err
-		}
-		rr, err := er.GetArrowRecordReader()
-		if err != nil {
-			return err
-		}
-		crr.rrs[i] = rr
-		crr.closers[i] = func() {
-			rr.Release()
-			er.Close()
-			reader.Close()
-		}
-	}
-	return nil
-}
-
-func (crr *compositeRecordReader) Next() error {
-	if crr.rrs == nil {
-		if crr.blobs == nil || len(crr.blobs) == 0 {
-			return io.EOF
-		}
-		crr.rrs = make([]array.RecordReader, len(crr.blobs))
-		crr.closers = make([]func(), len(crr.blobs))
-		crr.blobPos = -1
-		crr.fields = make([]FieldID, len(crr.rrs))
-		crr.r = compositeRecord{
-			recs:   make(map[FieldID]arrow.Record, len(crr.rrs)),
-			schema: make(map[FieldID]schemapb.DataType, len(crr.rrs)),
-		}
-		if err := crr.iterateNextBatch(); err != nil {
-			return err
-		}
-	}
-
-	composeRecord := func() bool {
-		for i, rr := range crr.rrs {
-			if ok := rr.Next(); !ok {
-				return false
-			}
-			// compose record
-			crr.r.recs[crr.fields[i]] = rr.Record()
-		}
-		return true
-	}
-
-	// Try compose records
-	if ok := composeRecord(); !ok {
-		// If failed the first time, try iterate next batch (blob), the error may be io.EOF
-		if err := crr.iterateNextBatch(); err != nil {
-			return err
-		}
-		// If iterate next batch success, try compose again
-		if ok := composeRecord(); !ok {
-			// If the next blob is empty, return io.EOF (it's rare).
-			return io.EOF
-		}
-	}
-	return nil
-}
-
-func (crr *compositeRecordReader) Record() Record {
-	return &crr.r
-}
-
-func (crr *compositeRecordReader) Close() {
-	for _, close := range crr.closers {
-		if close != nil {
-			close()
-		}
-	}
 }
 
 type serdeEntry struct {
@@ -627,52 +510,6 @@ var serdeMap = func() map[schemapb.DataType]serdeEntry {
 	return m
 }()
 
-func parseBlobKey(bolbKey string) (colId FieldID, logId UniqueID) {
-	if _, _, _, colId, logId, ok := metautil.ParseInsertLogPath(bolbKey); ok {
-		return colId, logId
-	}
-	if colId, err := strconv.ParseInt(bolbKey, 10, 64); err == nil {
-		// data_codec.go generate single field id as blob key.
-		return colId, 0
-	}
-	return -1, -1
-}
-
-func newCompositeRecordReader(blobs []*Blob) (*compositeRecordReader, error) {
-	sort.Slice(blobs, func(i, j int) bool {
-		iCol, iLog := parseBlobKey(blobs[i].Key)
-		jCol, jLog := parseBlobKey(blobs[j].Key)
-
-		if iCol == jCol {
-			return iLog < jLog
-		}
-		return iCol < jCol
-	})
-
-	blobm := make([][]*Blob, 0)
-	var fieldId FieldID = -1
-	var currentCol []*Blob
-
-	for _, blob := range blobs {
-		colId, _ := parseBlobKey(blob.Key)
-		if colId != fieldId {
-			if currentCol != nil {
-				blobm = append(blobm, currentCol)
-			}
-			currentCol = make([]*Blob, 0)
-			fieldId = colId
-		}
-		currentCol = append(currentCol, blob)
-	}
-	if currentCol != nil {
-		blobm = append(blobm, currentCol)
-	}
-
-	return &compositeRecordReader{
-		blobs: blobm,
-	}, nil
-}
-
 type DeserializeReader[T any] struct {
 	rr           RecordReader
 	deserializer Deserializer[T]
@@ -720,56 +557,6 @@ func NewDeserializeReader[T any](rr RecordReader, deserializer Deserializer[T]) 
 		rr:           rr,
 		deserializer: deserializer,
 	}
-}
-
-func NewBinlogDeserializeReader(blobs []*Blob, PKfieldID UniqueID) (*DeserializeReader[*Value], error) {
-	reader, err := newCompositeRecordReader(blobs)
-	if err != nil {
-		return nil, err
-	}
-
-	return NewDeserializeReader(reader, func(r Record, v []*Value) error {
-		// Note: the return value `Value` is reused.
-		for i := 0; i < r.Len(); i++ {
-			value := v[i]
-			if value == nil {
-				value = &Value{}
-				m := make(map[FieldID]interface{}, len(r.Schema()))
-				value.Value = m
-				v[i] = value
-			}
-
-			m := value.Value.(map[FieldID]interface{})
-			for j, dt := range r.Schema() {
-				if r.Column(j).IsNull(i) {
-					m[j] = nil
-				} else {
-					d, ok := serdeMap[dt].deserialize(r.Column(j), i)
-					if ok {
-						m[j] = d // TODO: avoid memory copy here.
-					} else {
-						return errors.New(fmt.Sprintf("unexpected type %s", dt))
-					}
-				}
-			}
-
-			if _, ok := m[common.RowIDField]; !ok {
-				panic("no row id column found")
-			}
-			value.ID = m[common.RowIDField].(int64)
-			value.Timestamp = m[common.TimeStampField].(int64)
-
-			pk, err := GenPrimaryKeyByRawData(m[PKfieldID], r.Schema()[PKfieldID])
-			if err != nil {
-				return err
-			}
-
-			value.PK = pk
-			value.IsDeleted = false
-			value.Value = m
-		}
-		return nil
-	}), nil
 }
 
 var _ Record = (*selectiveRecord)(nil)
@@ -984,164 +771,4 @@ func newSimpleArrowRecord(r arrow.Record, schema map[FieldID]schemapb.DataType, 
 		schema:    schema,
 		field2Col: field2Col,
 	}
-}
-
-type BinlogStreamWriter struct {
-	collectionID UniqueID
-	partitionID  UniqueID
-	segmentID    UniqueID
-	fieldSchema  *schemapb.FieldSchema
-
-	memorySize int // To be updated on the fly
-
-	buf bytes.Buffer
-	rw  *singleFieldRecordWriter
-}
-
-func (bsw *BinlogStreamWriter) GetRecordWriter() (RecordWriter, error) {
-	if bsw.rw != nil {
-		return bsw.rw, nil
-	}
-
-	fid := bsw.fieldSchema.FieldID
-	dim, _ := typeutil.GetDim(bsw.fieldSchema)
-	rw, err := newSingleFieldRecordWriter(fid, arrow.Field{
-		Name:     strconv.Itoa(int(fid)),
-		Type:     serdeMap[bsw.fieldSchema.DataType].arrowType(int(dim)),
-		Nullable: true, // No nullable check here.
-	}, &bsw.buf)
-	if err != nil {
-		return nil, err
-	}
-	bsw.rw = rw
-	return rw, nil
-}
-
-func (bsw *BinlogStreamWriter) Finalize() (*Blob, error) {
-	if bsw.rw == nil {
-		return nil, io.ErrUnexpectedEOF
-	}
-	bsw.rw.Close()
-
-	var b bytes.Buffer
-	if err := bsw.writeBinlogHeaders(&b); err != nil {
-		return nil, err
-	}
-	if _, err := b.Write(bsw.buf.Bytes()); err != nil {
-		return nil, err
-	}
-	return &Blob{
-		Key:        strconv.Itoa(int(bsw.fieldSchema.FieldID)),
-		Value:      b.Bytes(),
-		RowNum:     int64(bsw.rw.numRows),
-		MemorySize: int64(bsw.memorySize),
-	}, nil
-}
-
-func (bsw *BinlogStreamWriter) writeBinlogHeaders(w io.Writer) error {
-	// Write magic number
-	if err := binary.Write(w, common.Endian, MagicNumber); err != nil {
-		return err
-	}
-	// Write descriptor
-	de := newDescriptorEvent()
-	de.PayloadDataType = bsw.fieldSchema.DataType
-	de.CollectionID = bsw.collectionID
-	de.PartitionID = bsw.partitionID
-	de.SegmentID = bsw.segmentID
-	de.FieldID = bsw.fieldSchema.FieldID
-	de.StartTimestamp = 0
-	de.EndTimestamp = 0
-	de.descriptorEventData.AddExtra(originalSizeKey, strconv.Itoa(bsw.memorySize)) // FIXME: enable original size
-	if err := de.Write(w); err != nil {
-		return err
-	}
-	// Write event header
-	eh := newEventHeader(InsertEventType)
-	// Write event data
-	ev := newInsertEventData()
-	ev.StartTimestamp = 1 // Fixme: enable start/end timestamp
-	ev.EndTimestamp = 1
-	eh.EventLength = int32(bsw.buf.Len()) + eh.GetMemoryUsageInBytes() + int32(binary.Size(ev))
-	// eh.NextPosition = eh.EventLength + w.Offset()
-	if err := eh.Write(w); err != nil {
-		return err
-	}
-	if err := ev.WriteEventData(w); err != nil {
-		return err
-	}
-	return nil
-}
-
-func NewBinlogStreamWriters(collectionID, partitionID, segmentID UniqueID,
-	schema []*schemapb.FieldSchema,
-) map[FieldID]*BinlogStreamWriter {
-	bws := make(map[FieldID]*BinlogStreamWriter, len(schema))
-	for _, f := range schema {
-		bws[f.FieldID] = &BinlogStreamWriter{
-			collectionID: collectionID,
-			partitionID:  partitionID,
-			segmentID:    segmentID,
-			fieldSchema:  f,
-		}
-	}
-	return bws
-}
-
-func NewBinlogSerializeWriter(schema *schemapb.CollectionSchema, partitionID, segmentID UniqueID,
-	writers map[FieldID]*BinlogStreamWriter, batchSize int,
-) (*SerializeWriter[*Value], error) {
-	rws := make(map[FieldID]RecordWriter, len(writers))
-	for fid := range writers {
-		w := writers[fid]
-		rw, err := w.GetRecordWriter()
-		if err != nil {
-			return nil, err
-		}
-		rws[fid] = rw
-	}
-	compositeRecordWriter := newCompositeRecordWriter(rws)
-	return NewSerializeRecordWriter[*Value](compositeRecordWriter, func(v []*Value) (Record, uint64, error) {
-		builders := make(map[FieldID]array.Builder, len(schema.Fields))
-		types := make(map[FieldID]schemapb.DataType, len(schema.Fields))
-		for _, f := range schema.Fields {
-			dim, _ := typeutil.GetDim(f)
-			builders[f.FieldID] = array.NewBuilder(memory.DefaultAllocator, serdeMap[f.DataType].arrowType(int(dim)))
-			types[f.FieldID] = f.DataType
-		}
-
-		var memorySize uint64
-		for _, vv := range v {
-			m := vv.Value.(map[FieldID]any)
-
-			for fid, e := range m {
-				typeEntry, ok := serdeMap[types[fid]]
-				if !ok {
-					panic("unknown type")
-				}
-				ok = typeEntry.serialize(builders[fid], e)
-				if !ok {
-					return nil, 0, errors.New(fmt.Sprintf("serialize error on type %s", types[fid]))
-				}
-				writers[fid].memorySize += int(typeEntry.sizeof(e))
-				memorySize += typeEntry.sizeof(e)
-			}
-		}
-		arrays := make([]arrow.Array, len(types))
-		fields := make([]arrow.Field, len(types))
-		field2Col := make(map[FieldID]int, len(types))
-		i := 0
-		for fid, builder := range builders {
-			arrays[i] = builder.NewArray()
-			builder.Release()
-			fields[i] = arrow.Field{
-				Name:     strconv.Itoa(int(fid)),
-				Type:     arrays[i].DataType(),
-				Nullable: true, // No nullable check here.
-			}
-			field2Col[fid] = i
-			i++
-		}
-		return newSimpleArrowRecord(array.NewRecord(arrow.NewSchema(fields, nil), arrays, int64(len(v))), types, field2Col), memorySize, nil
-	}, batchSize), nil
 }
